+++ front-end/src/components/ContractManagementContainer.js
