+++ front-end/src/components/ContractPermissions.js
@@ -1,13 +1,12 @@
 import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { ethers } from "ethers";
-import ContractContainer from "../ContractRequestContainer";
+import ContractPrecompilerContainer from "./ContractPrecompileContainer";
 
-export default function ContractPageTransactions() {
+const ContractPermissions = () => {
   const [contractAddresses, setContractAddresses] = useState([]);
-  const [blockchainExists, setBlockchainExists] = useState(true);
   const [onDeleteRefreshState, onDeleteRefresh] = useState(true);
-
+  const [blockchainExists, setBlockchainExists] = useState(true);
   const { id } = useParams();
 
   useEffect(() => {
@@ -15,7 +14,6 @@ export default function ContractPageTransactions() {
     if (!subnets) {
       setBlockchainExists(false);
     }
-
   }, [id]);
 
   useEffect(() => {
@@ -54,48 +52,43 @@ export default function ContractPageTransactions() {
     setContractAddresses(contractAddresses);
   };
 
-  if (!blockchainExists) {
-    return (
-      <div className="flex flex-col items-center">
-        <div className="flex justify-center">
-          <h1 className="text-3xl font-bold mb-2 mt-4">Contract Page</h1>
-        </div>
-        <div className="flex flex-col items-center">
-          <h2 className="text-xl font-bold mb-2">No such blockchain exists</h2>
-          <a href="/">Go back to home page</a>
-        </div>
-      </div>
-    );
-  }
-
   return (
     <div className="flex flex-col items-center">
-      <div className="flex justify-center">
-        <h1 className="text-3xl font-bold mb-2 mt-4">Contract Page</h1>
-      </div>
-      <form
-        onSubmit={(event) => {
-          event.preventDefault();
-          const contractAddress = event.target.elements.contractAddress.value;
-          saveContractAddress(contractAddress);
-        }}
-      >
-        <input
-          type="text"
-          name="contractAddress"
-          placeholder="Contract Address"
-        />
-        <button className="mt-1 mb-4">Add Contract</button>
-      </form>
-      <ul>
-        {contractAddresses.map((contractAddress) => (
-          <ContractContainer
-            key={contractAddress}
-            contractAddress={contractAddress}
-            onDeleteRefresh={onDeleteRefresh}
+      <div className="flex flex-col items-center">
+        <div className="flex justify-center">
+          <h1 className="text-xl font-bold mb-2 text-gray-700">
+            Set Blockchain Permissions
+          </h1>
+        </div>
+        <form
+          onSubmit={(event) => {
+            event.preventDefault();
+            const contractAddress = event.target.elements.contractAddress.value;
+            saveContractAddress(contractAddress);
+          }}
+        >
+
+          <input
+            type="string"
+            name="contractAddress"
+            placeholder="Contract Address"
+            className="border-2 px-2 py-1 rounded-md w-80"
           />
-        ))}
-      </ul>
+          <button className="mt-2 mb-5">Add Distribution</button>
+        </form>
+
+        <ul>
+          {contractAddresses.map((contractAddress) => (
+            <ContractPrecompilerContainer
+              key={contractAddress}
+              contractAddress={contractAddress}
+              onDeleteRefresh={onDeleteRefresh}
+            />
+          ))}
+        </ul>
+      </div>
     </div>
   );
-}
+};
+
+export default ContractPermissions;
