+++ front-end/src/components/pages/ContractTransactionPage.js
@@ -1,12 +1,13 @@
 import React, { useState, useEffect } from "react";
 import { useParams } from "react-router-dom";
 import { ethers } from "ethers";
-import ContractPrecompilerContainer from "./ContractPrecompileContainer";
+import ContractContainer from "../ContractTransactionContainer";
 
-const BlockchainPermission = () => {
+export default function ContractPageTransactions() {
   const [contractAddresses, setContractAddresses] = useState([]);
-  const [onDeleteRefreshState, onDeleteRefresh] = useState(true);
   const [blockchainExists, setBlockchainExists] = useState(true);
+  const [onDeleteRefreshState, onDeleteRefresh] = useState(true);
+
   const { id } = useParams();
 
   useEffect(() => {
@@ -14,6 +15,7 @@ const BlockchainPermission = () => {
     if (!subnets) {
       setBlockchainExists(false);
     }
+
   }, [id]);
 
   useEffect(() => {
@@ -52,43 +54,48 @@ const BlockchainPermission = () => {
     setContractAddresses(contractAddresses);
   };
 
-  return (
-    <div className="flex flex-col items-center">
+  if (!blockchainExists) {
+    return (
       <div className="flex flex-col items-center">
         <div className="flex justify-center">
-          <h1 className="text-xl font-bold mb-2 text-gray-700">
-            Set Blockchain Permissions
-          </h1>
+          <h1 className="text-3xl font-bold mb-2 mt-4">Contract Page</h1>
         </div>
-        <form
-          onSubmit={(event) => {
-            event.preventDefault();
-            const contractAddress = event.target.elements.contractAddress.value;
-            saveContractAddress(contractAddress);
-          }}
-        >
-
-          <input
-            type="string"
-            name="contractAddress"
-            placeholder="Contract Address"
-            className="border-2 px-2 py-1 rounded-md w-80"
-          />
-          <button className="mt-2 mb-5">Add Distribution</button>
-        </form>
-
-        <ul>
-          {contractAddresses.map((contractAddress) => (
-            <ContractPrecompilerContainer
-              key={contractAddress}
-              contractAddress={contractAddress}
-              onDeleteRefresh={onDeleteRefresh}
-            />
-          ))}
-        </ul>
+        <div className="flex flex-col items-center">
+          <h2 className="text-xl font-bold mb-2">No such blockchain exists</h2>
+          <a href="/">Go back to home page</a>
+        </div>
       </div>
+    );
+  }
+
+  return (
+    <div className="flex flex-col items-center">
+      <div className="flex justify-center">
+        <h1 className="text-3xl font-bold mb-2 mt-4">Contract Page</h1>
+      </div>
+      <form
+        onSubmit={(event) => {
+          event.preventDefault();
+          const contractAddress = event.target.elements.contractAddress.value;
+          saveContractAddress(contractAddress);
+        }}
+      >
+        <input
+          type="text"
+          name="contractAddress"
+          placeholder="Contract Address"
+        />
+        <button className="mt-1 mb-4">Add Contract</button>
+      </form>
+      <ul>
+        {contractAddresses.map((contractAddress) => (
+          <ContractContainer
+            key={contractAddress}
+            contractAddress={contractAddress}
+            onDeleteRefresh={onDeleteRefresh}
+          />
+        ))}
+      </ul>
     </div>
   );
-};
-
-export default BlockchainPermission;
+}
